oppgaver/backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
azure-cognitiveservices-vision-computervision==0.9.1
azure-cognitiveservices-vision-customvision==3.1.1
azure-common==1.1.28
azure-core==1.32.0
azure-data-tables==12.6.0
azure-mgmt-core==1.5.0
azure-storage-blob==12.24.0
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==44.0.0
distro==1.9.0
Flask==3.1.0
Flask-Cors==5.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.0
MarkupSafe==3.0.2
msrest==0.7.1
multidict==6.1.0
oauthlib==3.2.2
openai==1.56.2
pillow==11.0.0
propcache==0.2.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
yarl==1.18.3
